rename search->search_state and move ot own files
