+++ Makefile.autosetup
@@ -694,7 +694,7 @@ $(PWD)/parse:
 LIBPATTERN=	libpattern.a
 LIBPATTERNOBJS=	pattern/compile.o pattern/config.o pattern/dlg_pattern.o \
 		pattern/exec.o pattern/flags.o pattern/functions.o \
-		pattern/message.o pattern/pattern.o
+		pattern/message.o pattern/pattern.o pattern/search_state.o
 CLEANFILES+=	$(LIBPATTERN) $(LIBPATTERNOBJS)
 ALLOBJS+=	$(LIBPATTERNOBJS)
