+++ index/shared_data.c
@@ -31,9 +31,9 @@
 #include "mutt/lib.h"
 #include "email/lib.h"
 #include "core/lib.h"
-#include "pattern/lib.h"
 #include "shared_data.h"
 #include "lib.h"
+#include "pattern/lib.h"
 #include "mview.h"
 
 /**
@@ -294,7 +294,7 @@ void index_shared_data_free(struct MuttWindow *win, void **ptr)
   if (shared->email)
     notify_observer_remove(shared->email->notify, index_shared_email_observer, shared);
 
-  mutt_search_free(&shared->search);
+  search_state_free(&shared->search);
 
   FREE(ptr);
 }
@@ -309,7 +309,7 @@ struct IndexSharedData *index_shared_data_new(void)
 
   shared->notify = notify_new();
   shared->sub = NeoMutt->sub;
-  shared->search = mutt_search_new();
+  shared->search = search_state_new();
 
   mutt_debug(LL_NOTIFY, "NT_INDEX_ADD: %p\n", (void *) shared);
   notify_send(shared->notify, NT_INDEX, NT_INDEX_ADD, shared);
