+++ index/shared_data.h
@@ -42,7 +42,7 @@ struct IndexSharedData
   struct Email        *email;          ///< Currently selected Email
   size_t               email_seq;      ///< Sequence number of the current email
   struct Notify       *notify;         ///< Notifications: #NotifyIndex, #IndexSharedData
-  struct Search       *search;         ///< State of the current index search
+  struct SearchState  *search;         ///< State of the current index search
 };
 
 void                    index_shared_data_free(struct MuttWindow *win, void **ptr);
